Clean up unused methods
