+++ src/Analyzers/SetExplicitMockBehaviorAnalyzer.cs
@@ -1,5 +1,4 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using Microsoft.CodeAnalysis.Operations;
+﻿using Microsoft.CodeAnalysis.Operations;
 
 namespace Moq.Analyzers;
 
@@ -155,20 +154,3 @@ public class SetExplicitMockBehaviorAnalyzer : DiagnosticAnalyzer
     }
 }
 
-internal static class IOperationExtensions
-{
-    /// <summary>
-    /// Walks down consecutive conversion operations until an operand is reached that isn't a conversion operation.
-    /// </summary>
-    /// <param name="operation">The starting operation.</param>
-    /// <returns>The inner non conversion operation or the starting operation if it wasn't a conversion operation.</returns>
-    public static IOperation WalkDownConversion(this IOperation operation)
-    {
-        while (operation is IConversionOperation conversionOperation)
-        {
-            operation = conversionOperation.Operand;
-        }
-
-        return operation;
-    }
-}
