+++ src/Common/IOperationExtensions.cs
@@ -1,0 +1,21 @@
+﻿using Microsoft.CodeAnalysis.Operations;
+
+namespace Moq.Analyzers.Common;
+
+internal static class IOperationExtensions
+{
+    /// <summary>
+    /// Walks down consecutive conversion operations until an operand is reached that isn't a conversion operation.
+    /// </summary>
+    /// <param name="operation">The starting operation.</param>
+    /// <returns>The inner non conversion operation or the starting operation if it wasn't a conversion operation.</returns>
+    public static IOperation WalkDownConversion(this IOperation operation)
+    {
+        while (operation is IConversionOperation conversionOperation)
+        {
+            operation = conversionOperation.Operand;
+        }
+
+        return operation;
+    }
+}
